pyproject.toml
[tool.poetry]
name = "ecovista-back-end"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
annotated-types = "0.7.0"
asgiref = "3.8.1"
build = "1.2.2.post1"
cachecontrol = "0.14.1"
certifi = "2024.8.30"
cffi = "1.17.1"
charset-normalizer = "3.4.0"
cleo = "2.1.0"
click = "8.1.7"
colorama = "0.4.6"
contextlib2 = "21.6.0"
crashtest = "0.4.1"
cryptography = "44.0.0"
distlib = "0.3.9"
django = "5.1.4"
django-ninja = "1.3.0"
django-ninja-extra = "0.21.8"
django-ninja-jwt = "5.3.5"
dulwich = "0.21.7"
fastjsonschema = "2.21.1"
filelock = "3.16.1"
ghp-import = "2.1.0"
idna = "3.10"
injector = "0.22.0"
installer = "0.7.0"
jaraco-classes = "3.4.0"
jinja2 = "3.1.4"
keyring = "24.3.1"
markdown = "3.7"
markupsafe = "3.0.2"
mergedeep = "1.3.4"
mkdocs = "1.6.1"
mkdocs-get-deps = "0.2.0"
more-itertools = "10.5.0"
msgpack = "1.1.0"
packaging = "24.2"
pathspec = "0.12.1"
pexpect = "4.9.0"
pillow = "11.0.0"
pkginfo = "1.12.0"
platformdirs = "4.3.6"
poetry = "1.8.5"
poetry-core = "1.9.1"
poetry-plugin-export = "1.8.0"
psycopg2 = "2.9.10"
ptyprocess = "0.7.0"
pycparser = "2.22"
pydantic = "2.10.3"
pydantic-core = "2.27.1"
pyjwt = "2.10.1"
pyproject-hooks = "1.2.0"
python-dateutil = "2.9.0.post0"
pywin32-ctypes = "0.2.3"
pyyaml = "6.0.2"
pyyaml-env-tag = "0.1"
rapidfuzz = "3.10.1"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
shellingham = "1.5.4"
six = "1.17.0"
sqlparse = "0.5.2"
tomlkit = "0.13.2"
trove-classifiers = "2024.10.21.16"
typing-extensions = "4.12.2"
tzdata = "2024.2"
urllib3 = "2.2.3"
virtualenv = "20.28.0"
watchdog = "6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
